add csvreader to library and create dir for csv example
