+++ examples/read_csv_example/config.json
@@ -1,7 +1,7 @@
 {
-  "XMPP_SERVER": "your_xmpp_server",
-  "artifact_name": "your_artifact_name",
-  "agent_name": "your_agent_name",
+  "XMPP_SERVER": "158.42.141.179",
+  "artifact_name": "art_csv",
+  "agent_name": "ag1",
   "csv_file_path": "df.csv",
   "columns_to_read": ["A", "B", "T"],
   "time_column": "T",
